[irods/irods#3452] igetwild cleanups
Use bash -e

set -e is good practice because it means the script will exit if a
command fails, rather than carrying on regardless.

Quote variables everywhere in igetwild

It's good practice to always quote shell variables (consider, for
instance, what would happen if TMPDIR contained a space).

Remove tmpfiles on error in igetwild

Only remove the tmpfiles at the end of a successful
run, or when an error occurs.

Signed-off-by: Matthew Vernon <mv3@sanger.ac.uk>
Signed-off-by: Terrell Russell <terrellrussell@gmail.com>