[irods/irods#5479] Use pam_password scheme name in iinit
Due to the nature of iinit, certain code paths are taken for certain
auth plugins and not others. These are referenced directly by name in
the client application and not done in any sort of generic fashion. We
plan to implement a new iinit in the future with the iRODS client CLI,
so for now we will go along with the current implementation and directly
reference the pam_password plugin as one which should not display a
password prompt.