[NOFIXUP,SQUASH] see entire commit message
distro_distill
procps
service template
service template for pre-253

To be appended to squashed commit message:

This commit introduces `distro_distill.py`, a slightly updated version of the
`distro_info.py` module currently in the externals repo. The goal is to
eventually make a standalone package for `distro_distill.py`. The buildsystem
now uses this instead of the `distro` Python module or `lsb_release` to
populate the `IRODS_LINUX_DISTRIBUTION_*` variables. Downstream projects may
need to be updated accordingly. As this results in potential changes to cache
variables, a clean build may be desirable.

Also:
  - [8123] Package systemd service unit template
  - [8125] Add `procps` to recommends for server deb package