[NOFIXUP,SQUASH] commit message
Append to commit message of squashed commit:

This commit additionally introduces the header `irods_chrono.hpp` in
`lib/core` which contains a couple of funcitons related to timestamp
generation. The code for getting the monotonic timestamp for service manager
communication has enough in common with the timestamp string generation code
used by the logger that it made sense to consolidate them both into the same
header.